refactor: remove v2 suffix
