+++ localstack/services/events/provider.py
@@ -55,7 +55,7 @@ from localstack.aws.api.events import EventBus as ApiTypeEventBus
 from localstack.aws.api.events import Rule as ApiTypeRule
 from localstack.services.events.event_bus import EventBusService, EventBusServiceDict
 from localstack.services.events.event_ruler import matches_rule
-from localstack.services.events.models_v2 import (
+from localstack.services.events.models import (
     EventBus,
     EventBusDict,
     EventsStore,
@@ -67,7 +67,7 @@ from localstack.services.events.models_v2 import (
 )
 from localstack.services.events.rule import RuleService, RuleServiceDict
 from localstack.services.events.target import TargetSender, TargetSenderDict, TargetSenderFactory
-from localstack.services.events.utils import (
+from localstack.services.events.v1.utils import (
     InvalidEventPatternException as InternalInvalidEventPatternException,
 )
 from localstack.services.plugins import ServiceLifecycleHook
